+++ app/(routes)/transfer/layout.tsx
@@ -1,0 +1,13 @@
+import { TransferUseFormDataProvider } from '@/contexts/TransferUseContext';
+
+export default function transferLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <TransferUseFormDataProvider>{children}</TransferUseFormDataProvider>
+    </>
+  );
+}
